+++ android/src/main/kotlin/org/codemaker/coder_matthews_extensions/CoderMatthewsExtensionsPlugin.kt
@@ -1,0 +1,66 @@
+package org.codemaker.coder_matthews_extensions
+
+import android.app.Activity
+import android.content.Context
+import androidx.window.layout.WindowMetricsCalculator
+
+import io.flutter.embedding.engine.plugins.FlutterPlugin
+import io.flutter.embedding.engine.plugins.activity.ActivityAware
+import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding
+import io.flutter.plugin.common.MethodCall
+import io.flutter.plugin.common.MethodChannel
+import io.flutter.plugin.common.MethodChannel.MethodCallHandler
+import io.flutter.plugin.common.MethodChannel.Result
+
+/** CoderMatthewsExtensionsPlugin */
+class CoderMatthewsExtensionsPlugin: FlutterPlugin, MethodCallHandler, ActivityAware {
+  /// The MethodChannel that will the communication between Flutter and native Android
+  ///
+  /// This local reference serves to register the plugin with the Flutter Engine and unregister it
+  /// when the Flutter Engine is detached from the Activity
+  private lateinit var channel : MethodChannel
+  private lateinit var context: Context
+  private lateinit var activity: Activity
+  override fun onAttachedToEngine(flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) {
+    channel = MethodChannel(flutterPluginBinding.binaryMessenger, "coder_matthews_extensions")
+    channel.setMethodCallHandler(this)
+    context = flutterPluginBinding.applicationContext
+
+  }
+
+  override fun onMethodCall(call: MethodCall, result: Result) {
+    when(call.method) {
+      "getPlatformVersion" -> result.success("Android ${android.os.Build.VERSION.RELEASE}")
+      "isTablet" ->{
+        val metrics = WindowMetricsCalculator.getOrCreate()
+          .computeCurrentWindowMetrics(activity)
+        val widthDp = metrics.bounds.width() /
+               context.resources.displayMetrics.density
+        val heightDp = metrics.bounds.height() /
+                context.resources.displayMetrics.density
+        var larger = if(widthDp > heightDp) widthDp else heightDp
+        result.success(larger > 900f)
+      }
+      else -> result.notImplemented()
+    }
+  }
+
+  override fun onDetachedFromEngine(binding: FlutterPlugin.FlutterPluginBinding) {
+    channel.setMethodCallHandler(null)
+  }
+
+  override fun onAttachedToActivity(binding: ActivityPluginBinding) {
+    activity = binding.activity
+  }
+
+  override fun onDetachedFromActivityForConfigChanges() {
+  }
+
+  override fun onReattachedToActivityForConfigChanges(binding: ActivityPluginBinding) {
+    activity = binding.activity
+  }
+
+  override fun onDetachedFromActivity() {
+  }
+
+}