+++ example/android/app/src/main/kotlin/org/codemaker/coder_matthews_extensions_example/MainActivity.kt
@@ -1,0 +1,6 @@
+package org.codemaker.coder_matthews_extensions_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}